move files around a bit
